refactor(evals): keep batch evaluation router thin
